+++ main.rkt
@@ -9,6 +9,7 @@
          tex-code
          latex-code
          objc-code
+         haskell-code
          pascal-code
          plist-code
          csv-code
@@ -33,6 +34,7 @@
          texblock
          latexblock
          objcblock
+         haskellblock
          pascalblock
          plistblock
          csvblock
@@ -57,6 +59,7 @@
          texblock0
          latexblock0
          objcblock0
+         haskellblock0
          pascalblock0
          plistblock0
          csvblock0
@@ -92,6 +95,7 @@
 (define-syntax-rule (tex-code . rest) (api:tex-code . rest))
 (define-syntax-rule (latex-code . rest) (api:latex-code . rest))
 (define-syntax-rule (objc-code . rest) (api:objc-code . rest))
+(define-syntax-rule (haskell-code . rest) (api:haskell-code . rest))
 (define-syntax-rule (pascal-code . rest) (api:pascal-code . rest))
 (define-syntax-rule (plist-code . rest) (api:plist-code . rest))
 (define-syntax-rule (csv-code . rest) (api:csv-code . rest))
@@ -116,6 +120,7 @@
 (define-syntax-rule (texblock . rest) (api:texblock . rest))
 (define-syntax-rule (latexblock . rest) (api:latexblock . rest))
 (define-syntax-rule (objcblock . rest) (api:objcblock . rest))
+(define-syntax-rule (haskellblock . rest) (api:haskellblock . rest))
 (define-syntax-rule (pascalblock . rest) (api:pascalblock . rest))
 (define-syntax-rule (plistblock . rest) (api:plistblock . rest))
 (define-syntax-rule (csvblock . rest) (api:csvblock . rest))
@@ -140,6 +145,7 @@
 (define-syntax-rule (texblock0 . rest) (api:texblock0 . rest))
 (define-syntax-rule (latexblock0 . rest) (api:latexblock0 . rest))
 (define-syntax-rule (objcblock0 . rest) (api:objcblock0 . rest))
+(define-syntax-rule (haskellblock0 . rest) (api:haskellblock0 . rest))
 (define-syntax-rule (pascalblock0 . rest) (api:pascalblock0 . rest))
 (define-syntax-rule (plistblock0 . rest) (api:plistblock0 . rest))
 (define-syntax-rule (csvblock0 . rest) (api:csvblock0 . rest))